--- actions/load_models/__init__.py
@@ -1,4 +1,0 @@
-from actions.load_models.speaker_encoder import (
-    load_speaker_transformer_encoder,
-    load_speaker_lstm_encoder,
-)
