--- actions/load_models/speaker_encoder/__init__.py
@@ -1,4 +1,0 @@
-from actions.load_models.speaker_encoder.transformer import (
-    load_speaker_transformer_encoder,
-)
-from actions.load_models.speaker_encoder.lstm import load_speaker_lstm_encoder
