--- actions/speaker_comparisons/__init__.py
@@ -1,3 +1,0 @@
-from actions.speaker_comparisons.comparison import calculate_cosine_similarity
-from actions.speaker_comparisons.processor import preprocess_audio
-from actions.speaker_comparisons.test import test_lstm_speaker_verification, test_transformer_speaker_verification
