+++ src/core/utils/configs/audio/__init__.py
@@ -1,0 +1,17 @@
+class AudioConfig:
+    N_MELS = None
+    SAMPLE_RATE = None
+    FRAME_SHIFT = None
+    FRAME_LENGTH = None
+    HOP_LENGTH = None
+    WIN_LENGTH = None
+    N_FFT = None
+    FMIN = None
+    FMAX = None
+    ZERO_THRESHOLD = None
+    MIN_AMPLITUDE = None
+    MAX_AMPLITUDE = None
+    MIN_LEVEL_DB = None
+    REF_LEVEL_DB = None
+    NUM_FRAMES = None
+    SCALING_FACTOR = None
