+++ src/speaker_verification/routers/__init__.py
@@ -1,0 +1,1 @@
+from speaker_verification.routers.verifications import router
