--- utils/configs/audio/__init__.py
@@ -1,2 +1,0 @@
-from utils.configs.audio.base import AudioConfig
-from utils.configs.audio.speaker_encoder import SpeakerEncoderAudioConfig
