--- utils/configs/audio/base.py
@@ -1,2 +1,0 @@
-class AudioConfig:
-    pass
