--- utils/objects/__init__.py
@@ -1,2 +1,0 @@
-from utils.objects.speaker import Speaker
-from utils.objects.utterance import Utterance
